requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-environ==0.8.1
djangorestframework==3.12.4
drf-spectacular==0.21.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.2.1
numpy==1.21.4
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
timezonefinder==5.2.0
typing-extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
